manifest: remove support for cy8cproto_062s3_4343w
This board only has 512 KB of internal flash, so it's not possible
to store the primary application in internal flash. This prevents
the project from compiling successfully.

There was an attempt to resolve the build errors by moving the start
address of the primary app to external flash in
`flashmap/psoc62_overwrite_single_smif.json`, but this resulted
in a linker error:

ld: section .text LMA [0000000018000400,00000000180962c7] overlaps section .cy_xip LMA [0000000018000000,0000000018068a64]

In the future, we may revisit this idea of moving the primary app to
external flash. But for now, we will simply remove support for this
board.

Signed-off-by: Nick Miller <nick@golioth.io>